+++ config/settings.py
@@ -100,8 +100,9 @@ REST_FRAMEWORK = {
     "DEFAULT_RENDERER_CLASSES": [
         "osidb.renderers.OsidbRenderer",
     ],
-    "DEFAULT_PAGINATION_CLASS": "rest_framework.pagination.LimitOffsetPagination",
+    "DEFAULT_PAGINATION_CLASS": "osidb.pagination.HardLimitOffsetPagination",
     "PAGE_SIZE": 100,
+    "MAX_PAGE_SIZE": 500,
     "DEFAULT_SCHEMA_CLASS": "drf_spectacular.openapi.AutoSchema",
     "EXCEPTION_HANDLER": "osidb.exception_handlers.exception_handler",
 }